Modification de la vue listing pour afficher les resultat de recherche

+++ view/clients_list_view.php
@@ -7,9 +7,21 @@
             <div class="grey-mask"></div>
         </div>
         <div class="container p-5">
+            
+            <?php if (isset($offset) && $offset !== "") {
+    ?> 
             <h1 class="display-4">Liste des clients</h1>
+            <?php
+} else {
+        ?>
+                   <h1 class="display-4">Recherche <span class="text-primary">"<?= $_GET['query']; ?>"</span></h1>
+            <p class="lead">dans "<?= $_GET['row']; ?>"</p>
+            <?php
+    } ?>
         </div>
     </div>
+    <?php if (isset($offset) && $offset !== "") {
+        ?>
     <div>
         <form class="form-inline mb-4 pcf" method="post" action="/change-list-size/">
 
@@ -21,6 +33,8 @@
             </select>
         </form>
     </div>
+    <?php
+    } ?>
     <div class="table-responsive">
         <table class="table table-striped list-clients">
             <thead class="thead-dark">
@@ -33,32 +47,50 @@
             </thead>
             <tbody>
                 <?php
-        foreach ($results as $result) {
-            $url = generateRewritedUrl($result['idClient'], $result['RaisonSociale']);
-            echo '<tr>';
-            echo '<td>'. $result['RaisonSociale'] .'</td>';
-            echo '<td>'. $result['Telephone'].'</td>';
-            echo '<td>'. $result['MailClient'].'</td>';
-            echo '<td>';
-            echo '<div class="btn-hover">';
-            echo '<a href="/consulter-client/'. $url .'" class="btn btn-outline-success"><i class="fas fa-eye"></i></a> ';
-            echo '<a href="/modifier-client/'. $url .'" class="btn btn-outline-warning"><i class="fas fa-pencil-alt"></i></a> ';
-            echo '<a href="/supprimer-client/'. $url .'" class="btn btn-outline-danger" data-toggle="modal" data-name="'. $result['RaisonSociale'] .'" data-url="/supprimer-client/'. $url .'"><i class="fas fa-trash-alt"></i></a>';
-            echo '</div>';
-            echo '</td>';
-            echo '</tr>';
-        }
+                if (!$results) {
+                    echo '<tr class="h-50-r"><td colspan="4" class="text-center"><p class="h5">Aucune résultats concernant votre recherche</p></td></tr>';
+                } else {
+                    foreach ($results as $result) {
+                        $url = generateRewritedUrl($result['idClient'], $result['RaisonSociale']);
+
+                        if (isset($offset) && $offset !== "") {
+                            $rs = $result['RaisonSociale'];
+                            $tel = $result['Telephone'];
+                            $mail = $result['MailClient'];
+                        } else {
+                            $rs =($_GET['row'] === 'raison-sociale') ?spanAround($_GET['query'], $result['RaisonSociale']) : $result['RaisonSociale'];
+                            $tel = ($_GET['row'] === 'telephone') ? spanAround($_GET['query'], $result['Telephone']) : $result['Telephone'];
+                            $mail = ($_GET['row'] === 'email') ? spanAround($_GET['query'], $result['MailClient']) : $result['MailClient'];
+                        }
+                            
+                        echo '<tr>';
+                        echo '<td>'. $rs . '</td>';
+                        echo '<td>'. $tel .'</td>';
+                        echo '<td>'. $mail .'</td>';
+                        echo '<td>';
+                        echo '<div class="btn-hover">';
+                        echo '<a href="/consulter-client/'. $url .'" class="btn btn-outline-success"><i class="fas fa-eye"></i></a> ';
+                        echo '<a href="/modifier-client/'. $url .'" class="btn btn-outline-warning"><i class="fas fa-pencil-alt"></i></a> ';
+                        echo '<a href="/supprimer-client/'. $url .'" class="btn btn-outline-danger" data-toggle="modal" data-name="'. $result['RaisonSociale'] .'" data-url="/supprimer-client/'. $url .'"><i class="fas fa-trash-alt"></i></a>';
+                        echo '</div>';
+                        echo '</td>';
+                        echo '</tr>';
+                    }
+                }
         ?>
             </tbody>
         </table>
+        <?php if (isset($offset) && $offset !== "") {
+            ?>
         <div class="text-center">
             <?php
                 for ($i = 1; $i <= $nbrpages; $i++) {
                     $active = ($_GET['page'] == $i) ? 'active' : '';
                     echo '<a href="/liste-clients/'.$i.'/" role="button" class="btn btn-outline-secondary '.$active.'">'.$i.'</a> ';
-                }
-            ?>
+                } ?>
         </div>
+        <?php
+        } ?>
     </div>
 </div>
 <!-- Modal -->
